utils: Split downloading csv from rest of code.
