+++ utils/flipflop-generate.py
@@ -2,78 +2,21 @@
 
 import argparse
 import csv
+import hashlib
 import json
+import os.path
 import pprint
 import textwrap
 import urllib.request
 
-
-SHEETS = {
-    'Yosys': 0,
-    'iCE40': 2076254101,
-    'ECP5': 1996872500,
-    'XC7': 1006733907,
-}
-
-def get_goog_sheet(sheet):
-    data = urllib.request.urlopen(
-        "https://docs.google.com/spreadsheets/d/e/2PACX-1vRxRhk3fCxPMh0yUSXorkA_FCdqDguW0JNKV5-Q64Xyi_Q9bI9mCc_Vfaw6DeBHFnd9MKsRZy2SrCgP/pub?output=csv&gid={}&single=true".format(SHEETS[sheet]),
-    ).read().decode('utf-8')
-
-    flipflops = []
-    headers = []
-    for i, row in enumerate(csv.reader(data.splitlines(), dialect='excel')):
-        if i == 0:
-            assert not headers
-            current = ''
-            for i in range(0, len(row)):
-                if row[i]:
-                    current = row[i]
-                row[i] = current
-            headers = row
-            continue
-        elif i == 1:
-            for i in range(0, len(row)):
-                v = row[i]
-                try:
-                    v = v[:v.index(" /")]
-                except ValueError:
-                    pass
-                try:
-                    v = v[:v.index(" (")]
-                except ValueError:
-                    pass
-                headers[i] = (headers[i], v)
-            pprint.pprint(headers)
-            continue
-
-        ff = {}
-        for (a, b), v in zip(headers, row):
-            if not v:
-                continue
-
-            if not b:
-                assert a not in ff, "{} (with value {}) found in {}".format(
-                    (a, b), v, ff
-                )
-                ff[a] = v
-                continue
-
-            if a not in ff:
-                ff[a] = {}
-            assert b not in ff[a]
-            ff[a][b] = v
-
-        flipflops.append(ff)
-
-    return flipflops
+import flipflop
 
 
 def parse_args():
     parser = argparse.ArgumentParser(
         description='Generate flip flops and such.'
     )
-    parser.add_argument('--sheet', default='Yosys', choices=SHEETS.keys())
+    parser.add_argument('--sheet', default='Yosys', choices=flipflop.SHEETS.keys())
     args = parser.parse_args()
 
     return args
@@ -86,59 +29,6 @@ def truth_table(ff):
 
 def escape_verilog(s):
     return s.replace('$', '\\$')
-
-
-def check_flipflop(ff):
-    """Check the flipflop has a valid configuration."""
-    try:
-        # Check the required top level properties exist
-        for s in ['Name', 'SyncAsync', 'Signal Names', 'Polarity', 'Type',
-                  'Class']:
-            assert s in ff, "Missing required '{}' property!".format(s)
-
-        # Check all the named signals also have polarity values.
-        for s in ff['Signal Names']:
-            if s in ('Data Out', 'Data In'):
-                continue
-            assert s in ff['Polarity'
-                           ], "Signal {} doesn't have a polarity!".format(s)
-
-        # Check if it has a Set or Reset signals it has init values for them
-        for s in ['Set', 'Reset']:
-            if s not in ff['Signal Names']:
-                continue
-            assert 'Init Values' in ff, "Missing 'Init Values' but have {} signal!".format(
-                s
-            )
-            assert s in ff[
-                'Init Values'
-            ], "Signal {} exists but no init value for it!".format(s)
-
-        # Check output signal
-        assert 'Data Out' in ff['Signal Names'], "Missing 'Data Out' signal!"
-
-        # Check if it is a SR flip flop or latch, then it has both Set and Reset signals
-        if 'SR' in ff['Name']:
-            assert 'Set' in ff['Signal Names'
-                               ], "SR flip flop but no 'Set' signal!"
-            assert 'Reset' in ff['Signal Names'
-                                 ], "SR flip flop but no 'Reset' signal!"
-
-            assert ff['Init Values'][
-                'Set'] == '1', "With SR flip flop 'Set' init value should be 1"
-            assert ff['Init Values'][
-                'Reset'
-            ] == '0', "With SR flip flop 'Reset' init value should be 0"
-
-        # Check if it is a D flip flop or latch it has a Data In signal
-        if ff['Name'].startswith('D'):
-            assert 'Data In' in ff['Signal Names'], "Missing 'Data In' signal!"
-
-    except AssertionError as e:
-        # Add a pretty printed version of the flip flop to the exception to
-        # make it easy to see what is going on.
-        e.args = (e.args[0] + '\n\n' + pprint.pformat(ff), *e.args[1:])
-        raise
 
 
 def describe(ff):
@@ -209,8 +99,6 @@ value_polarity = {
 
 def verilog(ff):
     """Get string containing verilog code for flipflop."""
-
-    check_flipflop(ff)
 
     # Generate the ports info
     # --------------------------------------------------------------------
@@ -326,7 +214,7 @@ endmodule
 
 def main():
     args = parse_args()
-    flipflops = get_goog_sheet(args.sheet)
+    flipflops = flipflop.flipflops(args.sheet)
     for i, ff in sorted(enumerate(flipflops), key=lambda ff: ff[-1]['Name']):
         print(i + 3, "-" * 70)
         print(verilog(ff))
